Update [30/03/23]
Added:
> Added notification intermission messages for Wave Intermissions (time in between waves).
> Enemies Update function changed to no longer perform actions when Enemy entity is dead.
> Refactored CharacterEntity class functions and parameters to use floats instead of integers.
> Readded stopping distance to Enemy Prefab.
> Added an Interact Notification object to the Screen Space Canvas of the GameScene.
> Updated the Canvas to Scale with Screen size.
> Implemented Interact Notification functions to the NotificationHandler and refactored the class.
> Updated the Player Script for IsDead logic.
> Added an ItemHandler Script for Item logic.
> Implemented a distance check for Interactions for ItemHandler.
> Added an Interaction button [F] to the InputActions.
> Updated InputManager for Interaction logic.
> Implemented logic for interacting with an Item as a Player.
> Implemented ModifyParameters functions for the CharacterEntity class.
> MOVESPEED definement in PlayerController Script moved from Start to Awake.

What does this mean:
> A notification indicating the time in between waves is now displayed when relevant.
> Action logic for entities are now disabled when they or related entities are 'dead'.
> UI now scaled to the Screen size, and is able to display Interact Notifications when relevant.
> Basic Item logic has now been implemented and can be interacted with.
> CharacterEntities can now be affected by parameter changes, and this is reflected in real-time.

+++ Assets/_Scripts/CharacterEntity/Enemy/Enemy.cs
@@ -20,7 +20,7 @@ public class Enemy : CharacterEntity {
         
     }
 
-    public override void TakeDamage(int playerParamATK) {
+    public override void TakeDamage(float playerParamATK) {
 
         paramHP -= CalculateDamage(playerParamATK);
